+++ Math/Digits/Printing First and Last digits of a number.cpp
@@ -1,0 +1,63 @@
+#pragma warning(disable:4786)
+#pragma warning(disable:4996)
+#include<bits/stdc++.h>
+#include <ext/pb_ds/assoc_container.hpp>
+#include <ext/pb_ds/trie_policy.hpp>
+#include<ext/pb_ds/detail/standard_policies.hpp>
+#define pii pair<int,int>
+#define pll pair<long long ,long long>
+#define pli pair<long long , int>
+#define pil pair<int ,long long>
+#define pi acos(-1)
+#define pb push_back
+#define mkp make_pair
+#define all(a) a.begin(), a.end()
+#define IOS ios::sync_with_stdio(0); cin.tie(0); cout.tie(0);
+#define ll long long
+#define MAX 300005
+#define INF 0x3f3f3f3f
+template <class T> inline T bigmod(T p,T e,T M){ll ret = 1LL;for(; e > 0LL; e >>= 1LL){if(e & 1LL) ret = (ret * p) % M;p = (p * p) % M;}return (T)ret;}
+template <class T> inline T modinverse(T a,T M){return bigmod(a,M-2,M);}   // M is prime}
+
+using namespace std;
+using namespace __gnu_pbds;
+
+typedef trie<string,null_type,trie_string_access_traits<>,pat_trie_tag,trie_prefix_search_node_update>pref_trie;
+typedef tree<ll,null_type,less<ll>,rb_tree_tag,tree_order_statistics_node_update>ordered_set;
+typedef tree<pair<int, int>,null_type,less<pair<int, int>>,rb_tree_tag,tree_order_statistics_node_update> ordered_multiset;
+
+long long binpow(long long a, long long b, long long m) {
+    a %= m;
+    long long res = 1;
+    while (b > 0) {
+        if (b & 1)
+            res = res * a % m;
+        a = a * a % m;
+        b >>= 1;
+    }
+    return res;
+}
+
+
+int main(){
+    IOS
+    //freopen("output.txt","w",stdout);
+    int t;
+    cin>>t;
+
+    for(int i=0;i<t;i++){
+        ll n,k;
+        cin>>n>>k;
+        cout<<(int)((pow(10,fmod(k*log10(n),1)))*100)<<"...";
+        ll l = binpow(n,k,1000);
+        if(l<10){
+            cout<<"00"<<l<<"\n";
+        }
+        else if(l<100){
+            cout<<"0"<<l<<"\n";
+        }
+        else{
+            cout<<l<<"\n";
+        }
+    }
+}
