+++ Math/String & Integer/multiply a string with an int.cpp
